Enhance frame stacking and logging in `train.py` and `PPOAgent`
This commit introduces frame stacking in the `collect_trajectory` function of `train.py`, allowing the agent to utilize a history of the last four frames for improved state representation. The training loop has been updated to ensure that the next state is also stacked before being stored. Additionally, the `PPOAgent` class has been refactored to enhance the save and load methods, ensuring that all relevant agent states, including menu action data and training progress variables, are properly managed. Logging improvements have been made throughout to provide better insights into the training process and potential issues. These changes aim to enhance the agent's performance and adaptability during training.

+++ src/environment/autonomous_reward_system.py
@@ -48,8 +48,9 @@ class WorldModelCNN(nn.Module):
             width, height = getattr(config, 'resolution', (1920, 1080))
             # Scale down to match the expected input resolution for the model
             model_width, model_height = 320, 240  # Typical processing resolution
-            # Create a dummy input to compute output shape
-            dummy_input = torch.zeros(1, 3, model_height, model_width, device=self.device)
+            # Create a dummy input to compute output shape - MUST use the same dtype as the encoder
+            dummy_input = torch.zeros(1, 3, model_height, model_width, 
+                                     device=self.device, dtype=self.dtype)
             encoder_output = self.encoder(dummy_input)
             flattened_size = encoder_output.numel() // encoder_output.size(0)
